refactor: decouple executor from build context
